* time.c (rb_time_unmagnify_to_float): should cast.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27556 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ time.c
@@ -755,7 +755,7 @@ rb_time_unmagnify_to_float(wideval_t w)
         if (c * b == a) {
             return DBL2NUM((double)c);
         }
-        v = DBL2NUM(FIXWV2WINT(w));
+        v = DBL2NUM((double)FIXWV2WINT(w));
         return quo(v, DBL2NUM(TIME_SCALE));
     }
 #endif
